adflib: ADFLib is now 64-bit compilable

git-svn-id: https://svn.aros.org/svn/aros/trunk@45764 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ tools/adflib/adf_bitm.h
@@ -9,7 +9,6 @@
  */
 
 #include"adf_str.h"
-#include"prefix.h"
 
 RETCODE adfReadBitmapBlock(struct Volume*, SECTNUM nSect, struct bBitmapBlock*);
 RETCODE adfWriteBitmapBlock(struct Volume*, SECTNUM nSect, struct bBitmapBlock*);
@@ -18,7 +17,7 @@ RETCODE adfWriteBitmapExtBlock(struct Volume*, SECTNUM, struct bBitmapExtBlock* 
 
 SECTNUM adfGet1FreeBlock(struct Volume *vol);
 RETCODE adfUpdateBitmap(struct Volume *vol);
-PREFIX long adfCountFreeBlocks(struct Volume* vol);
+PREFIX uint32_t adfCountFreeBlocks(struct Volume* vol);
 RETCODE adfReadBitmap(struct Volume* , SECTNUM nBlock, struct bRootBlock* root);
 BOOL adfIsBlockFree(struct Volume* vol, SECTNUM nSect);
 void adfSetBlockFree(struct Volume* vol, SECTNUM nSect);
